+++ src/ATen/native/quantized/sycl/AffineQuantizerKernels.cpp
