+++ src/ATen/native/xpu/sycl/RecordStream.cpp
@@ -11,6 +11,9 @@
 namespace at::native {
 void record_stream_xpu(Tensor& self, c10::Stream stream) {
   struct c10::StreamData3 data = stream.pack3();
-  c10::xpu::XPUCachingAllocator::recordStream(self.storage().data_ptr(), at::xpu::XPUStream::unpack3(data.stream_id, data.device_index, data.device_type));
+  c10::xpu::XPUCachingAllocator::recordStream(
+      self.storage().data_ptr(),
+      at::xpu::XPUStream::unpack3(
+          data.stream_id, data.device_index, data.device_type));
 }
-}  // namespace at::native
+} // namespace at::native
